cleanup `VarElem` interface; also weight->base-weight
