+++ crates/typst-library/src/math/root.rs
@@ -130,7 +130,7 @@ fn layout(
         line_pos,
         FrameItem::Shape(
             Geometry::Line(Point::with_x(radicand.width())).stroked(Stroke {
-                paint: ctx.default_var_fill(),
+                paint: var_fill(None, ctx.styles()),
                 thickness,
                 ..Stroke::default()
             }),
